Only need to hold up to 24

+++ src/days/mod.rs
@@ -14,7 +14,7 @@ pub trait ChristmasDay {
     }
 }
 
-pub fn run(day: i32, data: &String) -> (String, String){
+pub fn run(day: i8, data: &String) -> (String, String){
     let fnct: Option<Box<ChristmasDay>> = match day {
         1 => Some(Box::new(day1::Day1)),
         2 => Some(Box::new(day2::Day2)),
